requirements.txt
googleapis-common-protos
grpcio
protobuf
pyaml